# Use "All" page of requested entity as starting url for session