--- a/airgun/entities/architecture.py
+++ b/airgun/entities/architecture.py
@@ -10,6 +10,7 @@ from airgun.views.architecture import (
 
 
 class ArchitectureEntity(BaseEntity):
+    endpoint_path = '/architectures'
 
     def create(self, values):
         """Create new architecture entity"""