--- a/airgun/entities/configgroup.py
+++ b/airgun/entities/configgroup.py
@@ -10,6 +10,7 @@ from airgun.views.configgroup import (
 
 
 class ConfigGroupEntity(BaseEntity):
+    endpoint_path = '/config_groups'
 
     def create(self, values):
         """Create new config group"""