--- a/airgun/entities/contenthost.py
+++ b/airgun/entities/contenthost.py
@@ -12,6 +12,7 @@ from airgun.views.job_invocation import (
 
 
 class ContentHostEntity(BaseEntity):
+    endpoint_path = '/content_hosts'
 
     def delete(self, entity_name):
         """Delete existing content host"""