--- a/airgun/entities/host.py
+++ b/airgun/entities/host.py
@@ -23,6 +23,7 @@ from airgun.views.host import (
 
 
 class HostEntity(BaseEntity):
+    endpoint_path = '/hosts'
 
     HELPER_CLASS = HostHelper