--- a/airgun/entities/provisioning_template.py
+++ b/airgun/entities/provisioning_template.py
@@ -11,6 +11,7 @@ from airgun.views.provisioning_template import (
 
 
 class ProvisioningTemplateEntity(BaseEntity):
+    endpoint_path = '/provisioning_templates'
 
     def create(self, values):
         """Create new provisioning template"""