--- a/airgun/entities/rhai/manage.py
+++ b/airgun/entities/rhai/manage.py
@@ -5,6 +5,7 @@ from airgun.views.rhai import ManageDetailsView
 
 
 class ManageEntity(BaseEntity):
+    endpoint_path = '/redhat_access/insights/manage'
 
     def _toggle_service(self, state):
         view = self.navigate_to(self, "Details")