--- a/airgun/entities/smart_class_parameter.py
+++ b/airgun/entities/smart_class_parameter.py
@@ -7,6 +7,7 @@ from airgun.views.smart_class_parameter import (
 
 
 class SmartClassParameterEntity(BaseEntity):
+    endpoint_path = '/puppetclass_lookup_keys'
 
     def search(self, value):
         """Search for smart class parameter entity and return table row that