--- a/airgun/entities/task.py
+++ b/airgun/entities/task.py
@@ -4,6 +4,7 @@ from airgun.views.task import TaskDetailsView, TasksView
 
 
 class TaskEntity(BaseEntity):
+    endpoint_path = '/foreman_tasks/tasks'
 
     def search(self, value):
         """Search for specific task"""