--- a/airgun/session.py
+++ b/airgun/session.py
@@ -251,6 +251,7 @@ class Session(object):
         try:
             selenium_browser = self._factory.get_browser()
             self.browser = AirgunBrowser(selenium_browser, self)
+            LOGGER.info('Setting initial URL to {url}')
 
             self.browser.url = url