This is Airgun-only re-implementation of idea presented by @omkarkhatavkar in https://github.com/SatelliteQE/robottelo/pull/7442

The main goal here is to use entity "All items" URL as starting point for session. This way list of all entities will be visible right after logging in (instead of displaying dashboard, which we hardly ever need).

This is implemented by deferring execution of browser initialization code until entity is requested. It has two effects:
1. We can let each entity define URL fragment on their own. Web-specific data is kept within Airgun, which is in line of our design principles. As a result, this change is completely transparent to Robottelo - no test must be changed to benefit from this PR.
2. If you happen to use `Session` context manager, but never actually interact with it, browser will not be initialized. This will save time during test development and allows us to put more code inside session context (as there is no performance penalty for that).

Test case:
```bash
$ pytest -s --durations=0 --count 3 -k end_to_end tests/foreman/ui/test_lifecycleenvironment.py
```

Before change:
```
=================================================================================== slowest test durations ====================================================================================
73.52s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[1-3]
71.91s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[2-3]
70.26s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[3-3]
6.75s setup    tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[1-3]
0.24s teardown tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[3-3]
```

After change:
```
=================================================================================== slowest test durations ====================================================================================
61.88s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[3-3]
61.47s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[1-3]
60.53s call     tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[2-3]
4.36s setup    tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[1-3]
0.26s teardown tests/foreman/ui/test_lifecycleenvironment.py::test_positive_end_to_end[3-3]
```

So, with that PR, we can expect individual test execution time to be reduced by **8 to 13 seconds**.

I am also running entire UI Robottelo suite, but this takes forever.