@mirekdlugosz 2019-11-18
With bd1653b019b9f3d74e0c77e2be67d366f7d3bdd6, both of following idioms are supported:
```python
with session:
   # do something - starting URL will be read from requested entity
```
```python
with session(url="url")
   # do something - starting URL will be $HOSTNAME/url, i.e. provided by test and not entity
```
We initially discussed `session('url')` format, but I decided to generalize this to other arguments supported by `__init__`. So you can also use it to login as specific user (`session(user_name, user_pass)`), or pass session cookie obtained in test (`session(session_cookie=cookie))`).

I am very much interested on what you think about this API.