@omkarkhatavkar 2019-11-18
@mirekdlugosz overall looks nice, we are adding many ways access session makes more sense to me. I have one doubt here If we use plain `with session:` it will still navigate entity directly instead navigating from the menubar. For this, we can have the first test from each entity to be using plain session.   

 In other cases with `with session('entity/create')` will navigate to any required page.