@mirekdlugosz 2019-11-18
> I have one doubt here If we use plain `with session:` it will still navigate entity directly instead navigating from the menubar. 

With plain `with session:`, it will be:
```
All page for entity --(use button)-> New entity page
```

I have checked on `test_positive_create_with_cv` from `tests/foreman/ui/test_activationkey.py`:

```
# with session:
$ pytest -v tests/foreman/ui/test_activationkey.py -k 'test_positive_create_with_cv[1]'
==================================================================== 1 passed, 42 deselected, 1 warnings in 97.66 seconds =====================================================================
```

```
# with session(url='/activation_keys/new')
$ pytest -v tests/foreman/ui/test_activationkey.py -k 'test_positive_create_with_cv[1]'
==================================================================== 1 passed, 42 deselected, 1 warnings in 95.05 seconds =====================================================================
```

Of course these are only single runs for each.

So yes, it will reduce execution time a little further if we change robottelo to set initial URL to that of "new entity" page. But that's for another PR, as it must be done in robottelo and not airgun.