@omkarkhatavkar 2019-11-18
> So yes, it will reduce execution time a little further if we change robottelo to set initial URL to that of "new entity" page. But that's for another PR, as it must be done in robottelo and not airgun.

Absolutely, thanks for examples